\name{ExPD2D-package}
\alias{ExPD2D}
\alias{ExPD2D}
\docType{package}
\title{Exact computation of bivariate projection depth based on FORTRAN code}
\description{
Projection depth is one of favorite data depth among all its competitors. But like all depth functions, compute data depth is a very challenging problem. This package computes exact projection depth based on FORTRAN code, but now one can utilize the package in R to get the exact projection depth of data points.
}
\details{
\tabular{ll}{
Package: \tab ExPD2D\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2009-05-06\cr
License: \tab GPL version 2\cr
LazyLoad: \tab yes\cr
}
}
\author{Yijun Zuo, Xiangyang Ye}

\references{
Yijun Zuo, Projection-based Depth Functions and Associated Medians. \emph{The Annals of Statistics} 2003; Vol 31, No. 5 1460-1490\cr

Yijun zuo, Exactly Compute Bivariate Projection Depth and Stahel-Donoho Estimator. Manuscript, 2009
}

\keyword{ package }

