\name{pds}
\alias{pds}
\docType{data}
\title{Sample Data Frame}
\description{
A bivariate data set with 19 observations.
}
\usage{data(pds)}
\format{
  A data frame with 19 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{A numeric vector (x coordinates of the given data)}
    \item{\code{y}}{A numeric vector (y coordinates of the given data}
  }
}

\examples{
data(pds)
pds
}
\keyword{datasets}
